requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
groq>=0.4.2
langchain>=0.1.0
langchain-groq>=0.0.1
pydantic>=2.5.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pypdf2>=3.0.1
pillow>=10.1.0
pytesseract>=0.3.10
opencv-python>=4.8.1.78
numpy>=1.24.3
pandas>=2.1.4
jinja2>=3.1.2
aiofiles>=23.2.1
